Cherrypicked Commit - Add log component
The log component can write to a file or to stdout/stderr. It is not
threadsafe, and borrows the mutex and file descriptor used. It could
theoretically be used to write to other file descriptors without much
work, but since logging as written is a blocking operation this would
be risky.

It supports these log levels:
  - off
  - error
  - warn
  - info
  - debug

The profiler uses the component to build a thread-safe-ish abstraction
for a global profiler logger. The profiler uses the log levels to mean
roughly the following:
  - off: nothing will be logged
  - error: an error that will likely stop the profiler, such as failing
    to spawn a worker thread.
  - warn: an error that will likely not stop the profiler, such as
    an HTTP 5xx error on upload. The profile will be dropped, but the
    profiler will continue to gather samples and upload profiles.
  - info: a notable event that occurs infrequently, such as profiles
    being uploaded (happens once about every minute). If something
    happens once per request, that would be too noisy for this level.
  - debug: information intended for developers to understand internal
    order of events. May be very noisy.