Cherrypicked Commit - Add queue and channel components (#1388)
* Add queue and channel components

The queue is bounded and NOT thread-safe.

The channel is a bounded, multiple-producer, single-consumer channel
suitable for inter-thread communication.

The profiler will use a channel to send stack samples to another thread
to be aggregated.

* Document uv_cond_timedwait's unlocking/relocking

Multiple people asked about the lock being held while it waits, which
isn't the case, so let's document that in the source.