Cherrypicked Commit - Add time component
The time component accounts for differences in cpu-time calls in
different platforms. In the past it was also used for monotonic and
system time, which is why it is called "time" instead of "cpu-time".

There are no tests for this component -- I could not think of any tests
that would be meaningful.