Cherrypicked Commit - Properly report error for artisan command in Laravel 7+ (#1456)
* Properly report error for artisan command in Laravel 7+

Laravel 7 introduced a change in the underlying version of the Symfony libraries.

One of the change made in the Symfony Console lib, was to deprecate the `Symfony\Component\Console\Application#renderException` method
in favor of the `renderThrowable` method.

Datadog instrumentation does not take that change in account.
Effectively, if a Laravel artisan trace fails the root span will be reported as an error.

With this fix, it will be possible to track failing artisan commands in Laravel 7+.

* Add CLI test for Laravel 8_X

* moved from 'legacy' userland spans to internal spans

Co-authored-by: Luca Abbati <luca.abbati@gmail.com>

Co-authored-by: Luca Abbati <luca.abbati@gmail.com>